Deleted old ExcelUtility.java file and added setCellData method using try-with-resources.
